refactor method names in tests
Resolves:
BI-13454
